pyproject.toml
[project]
name = "language-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs>=2.9.2",
    "fastapi>=0.116.1",
    "ffmpeg>=1.4",
    "google-generativeai>=0.8.5",
    "openai>=1.99.6",
    "openai-whisper>=20250625",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
    "whisper>=1.1.10",
]
